requirements.txt
# Latest versions (2024/2025) - Updated requirements

# Core LangChain (0.2+)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.1.0
langchain_mcp_adapters

# Vector DB
chromadb>=0.5.0

# Embeddings and LLM
openai>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Optional: For local LLM support
ollama>=0.1.0

# API Server (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.24.0
black>=23.0.0
mypy==1.11.0
